Add ContactInformation model
We need to being able to retrieve, create, update and delete contact
information, hence here we create the model for just that.

The model shall include:
- Social security number
- Email address
- Phone number

Additionally we add an Id to identify each record.

This model will serve as a DTO and a domain model (db) as we do not have
time to do this properly and separating our endpoint logic and
the database.