Create ContactInformationController and Mock data class
We are going to do REST operations on contact information, so here
we prepare for doing so by adding the ContactInformationController
and MockContactInformationData class, which will help us doing the
operations and storing the data. We add MockContactInformationData
as a singleton as we want the data to remain the same during the
lifetime of the application, so that we can run REST operations on
data and still being able to see the result.

The MockContactInformationData methods are not yet implemented but
will be later on.

While at it, remove not used usings.