Add AddContactInfoprmation POST api endpoint
Add a api endpoint on which we can do http post and add a new
contact information record.

When the operation is done, return the newly added record with a
201 created response.

We also add a TODO for adding validation on the model as it has some
requirements, we handle that later.