Add ContactInformation validation
We have to ensure that the models passed to the API endpoints are
valid, we do this by using data annotations.

The social security number is an long, hence we cannot use the existing
string annotations and have to create our own, Lenght, which
takes a min and max length value, as a requirement is that this
number should be between 10 and 12.

Int cannot store as big numbers 10-12 can be, hence we go
with Long.