Change ContactsController route setup
Instead of having to state the api/[controller] each time on each
method, we instead put that on the controller, and only specify what
is unique per method. While at it, add a comment so it is clear what
the endpoint is.

Also, I decide to specify the controller name as we do not want to
rely on the controller.cs name, as we might want to change that later
but we do not want to change the controller route.