Make SocialSecurityNumber property nullable
We want the DTO model to have the SocialSecurityNumber as nullable
to be able to do proper validation on the required attribute.

The domain/db has this as not nullable hence this is not good and
should have been a dto class and a domain class so we can separate
this, but due to time, I wont.