Use CreatedAtRoute instead of depending on httpcontext
It is much cleaner to use CreatedAtRoute using the method names
which also will later on indicate to the developers if they change
the name of the method, some other place is depending on it.

Also add a comment about that we really should have had a dto model
here instead of the same model for the endpoint as we have for updating
the data, exposing the id which we do not want to.