Add UpdateContactInformation endpoint
We add a http patch endpoint for doing updates on our ContactInformation
datas, If success we return an Ok response and the object.

A lot of manual work here due to that we do not have all the structure
in place in this API, as of yet. With proper handling with db and DTO models.