Change Patch to Put
As we do not have DTO classes, the validation is wrong for a patch,
as all properties should be "not required" so that the client can
for example only update one property.

Hence we instead do a HTTP put, where we update the full object.

We decide that we do not do a upsert here, meaning, if the record is
missing, we do not insert a new record.

While at it, we noticed that we where passing the wrong ID, another
issue with not having DTO classes.