Add nuget AutoMapper
We use AutoMapper to map our DTOs to our db models. This way we
do it really easy and cleanly, this package serves really well when
the models are big.

Here we also setup AutoMapper and add our Profiler and add
the ContactInformationReadDto which we later on will use in our get methods.

While at it, also set it up in the ContactsController.

Issue: #3