Use and Create ContactInformationCreateDto
Here we create ContactInformationCreateDto to be used as the object
the client passes to create new ContactInformation objects, where
we omit the id property, as that is an internal primary key of the
DB model, which gets automatically added.

This model we should have the data annotations on so that we enforce
the required/non required and certain type of datas, such as email
and phone number.

Issue: #3