Create ContactInformationUpsertDto and us it in HTTP put
Change the model name ContactInformationCreateDto to
ContactInformationUpsertDto as the model is identical for update and
insert of ContactInformation.

Then also use it in our HTTP put method where it holds all the
properties the client should be able to update, and in code we
safely map it to our DB model ContactInformation.

Then return a ContactInformationReadDto for the user to be able to
get an awknowledge on what the end result is.

Whilst at it, remove a not used using statement.

Issue: #3